requirements.txt
fastapi
uvicorn
pyswisseph
pytz
pydantic
pydantic-settings
httpx
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
geopy
yfinance==0.2.50
timezonefinder
langchain-community
langchain
langchain-huggingface
pypdf
bcrypt==3.2.2
alembic
argon2-cffi
chromadb
email-validator
fastapi
geopy
google-generativeai
httpx
langchain
langchain-community
openai
pandas
passlib
pydantic
pydantic-settings
pypdf
pyswisseph
python-dotenv
python-jose
python-multipart
pytz
requests
sqlalchemy
timezonefinder
uvicorn
vedastro
yfinance
openai
google-genai
